modified security filter
