+++ Instructor-Application/src/main/java/com/cognixia/jump/exception/MyErrorController.java
@@ -1,0 +1,15 @@
+package com.cognixia.jump.exception;
+
+import org.springframework.boot.web.servlet.error.ErrorController;
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class MyErrorController implements ErrorController {
+
+  @GetMapping("/error")
+  public String handleError() {
+    // do something like logging
+    return "error";
+  }
+}
