+++ Instructor-Application/src/main/resources/application-prod.properties
@@ -1,0 +1,18 @@
+#Properties for Personal / Local Development
+
+# Hibernate generate DB Schemas from Spring Boot Models
+spring.jpa.hibernate.ddl-auto=update
+
+# Development Features to see SQL Commands from Hibernate in Console
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+springdoc.swagger-ui.path=/openapi.html
+# Windows info for MySQL login Local
+spring.datasource.url=jdbc:mysql://jump-java-oct-23-dev-db.cpp46bwoaqxm.us-east-1.rds.amazonaws.com:3306/spring_db
+spring.datasource.username=root
+spring.datasource.password=rootroot
+
+# Mac/Linux info for MySQL login
+#spring.datasource.url=jdbc:mysql://localhost:3306/spring_db?serverTimezone=EST5EDT
+#spring.datasource.username=root
+#spring.datasource.password=Root@123
