+++ Instructor-Application/src/main/resources/application.properties
@@ -1,19 +1,5 @@
-# if port 8080 is in use, can change our port number using the line below
-#server.port=8090
+# Default Properties that loads for all Profiles
+# Not much needed here because our profiles have some distinct behaviors
 
-# Mac/Linux info for MySQL login
-spring.jpa.hibernate.ddl-auto=update
-spring.datasource.url=jdbc:mysql://localhost:3306/spring_db
-spring.datasource.username=root
-spring.datasource.password=root
-
-# Windows info for MySQL login
-#spring.jpa.hibernate.ddl-auto=update
-#spring.datasource.url=jdbc:mysql://localhost:3306/spring_db
-#spring.datasource.username=root
-#spring.datasource.password=root
-
-# Shows logs for the queries hibernate uses for our interactions with our database
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-springdoc.swagger-ui.path=/openapi.html
+# set default profile to dev which has localhost configurations
+spring.profiles.active=dev
