--- instructor-application-frontend/src/App.css
@@ -1,105 +1,0 @@
-.head {
-  background-color: black;
-  height: 150px;
-  color: white;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.head-img {
-  width: 400px;
-}
-.login-form {
-  display: flex;
-  align-items: center;
-  /* padding: 200px; */
-  margin: 30px;
-  padding-left: 200px;
-  /* padding-top: 120px; */
-  /* background-color: aqua; */
-}
-.border {
-  background-color: black;
-  width: 300px;
-  height: 250px;
-  /* margin: 20px; */
-  padding: 25px 25px;
-}
-.login-label {
-  font-size: 2rem;
-  color: aliceblue;
-}
-.sign-up-label {
-  font-size: 1rem;
-  color: aliceblue;
-  gap: 10px;
-}
-.sign-up-input {
-  /* padding-left: 10px; */
-  /* padding-right: 10px; */
-}
-.login-input {
-  width: 200px;
-  height: 30px;
-}
-.btn {
-  margin: 5px;
-  /* padding-top: 50px; */
-}
-.grid-sepration {
-  display: grid;
-  grid-template-rows: 1fr; /* Single row */
-  grid-template-columns: 1fr 1fr; /* Two columns with equal width */
-  gap: 10px;
-}
-.collabera {
-  background-color: black;
-  width: 300px;
-  height: 270px;
-}
-.right {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 10px;
-}
-.list-spacing {
-  display: flex;
-  flex-direction: row;
-}
-.grid-container {
-  display: grid;
-  grid-template-columns: repeat(5, 1fr);
-  gap: 20px;
-  /* padding-top: 10px; */
-  margin: 20px;
-}
-.dash-title {
-  font-size: 5rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-.list {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 2rem;
-  padding-left: 500px;
-}
-.first-line {
-  display: flex;
-  justify-content: space-between;
-}
-.fnamespace {
-  padding-left: 30px;
-}
-.lnamespace {
-  padding-left: 32px;
-}
-.unamespace {
-  padding-left: 35px;
-}
-.passwordspace {
-  padding-left: 39px;
-}
